+++ R/exampleIsoscape.R
@@ -1,15 +1,15 @@
 #' Example isoscape data
 #'
 #'
-#' @format A dataframe containing 21600 observations, derived from a cropped raster object.
+#' @format A dataframe containing 1800 observations, derived from a cropped raster object.
 #'
 #' @references Bowen G. J., West J.B., Miller C. C., Zhao L. and Zhang T. (2018) IsoMAP: Isoscapes Modeling, Analysis and Prediction (version 1.0). IsoMAP job 66100, Caitlin J. Campbell. Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project.
 #' (\href{http://isomap.org}{http://isomap.org}).
 #'
-#' @source \href{https://isomap.rcac.purdue.edu/isomap/fitting/explore?jobseq=66085&caseName=Annual_NoAm_H_1980_2010&userName=caitjcampbell}{IsoMAP Job 66100}
+#' @source \href{https://isomap.rcac.purdue.edu/isomap/DownloadMetadata/download?content=tarball&type=raster&jobseq=66100&username=caitjcampbell&casename=Annual_NoAm_Map_H_1980_2010}{Download IsoMAP Job Map 66100}
 #'
 #'@examples
 #'data(isoscape)
 #'iso <- rasterFromXYZ(isoscape)
-#'
+
 "isoscape"