+++ R/exampleIsoscape_sd.R
@@ -1,0 +1,15 @@
+#' Example isoscape standard deviation data
+#'
+#'
+#' @format A dataframe containing 1800 observations, derived from a cropped raster object.
+#'
+#' @references Bowen G. J., West J.B., Miller C. C., Zhao L. and Zhang T. (2018) IsoMAP: Isoscapes Modeling, Analysis and Prediction (version 1.0). IsoMAP job 66100, Caitlin J. Campbell. Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project.
+#' (\href{http://isomap.org}{http://isomap.org}).
+#'
+#' @source \href{https://isomap.rcac.purdue.edu/isomap/DownloadMetadata/download?content=tarball&type=raster&jobseq=66100&username=caitjcampbell&casename=Annual_NoAm_Map_H_1980_2010}{Download IsoMAP Job Map 66100}
+#'
+#'@examples
+#'data(isoscape_sd)
+#'iso_sd <- rasterFromXYZ(isoscape_sd)
+
+"isoscape_sd"